+++ src/components/detail/DetailPoll.style.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const Wrap = styled.div`
+  width: 100%;
+`;
+
+export { Wrap }
