+++ src/pages/board/Board.jsx
@@ -89,6 +89,7 @@ const Board = () => {
   );
 };
 
+// for useRef
 const PostCardList = styled.div`
   margin-top: 10rem;
   height: 45.28rem;
